feat(frontend): center big button to create first satellite
